Fixed step over debugging.
The step over implementation specifically looked for call expressions to determine whether the step over was matched. This missed statements that had a call expression embedded, e.g. an assignment that assigned the result of a call expression. The algorithm has been changed to be triggered by an actual call instead.