Replaced source string with a class.
The source string on Location has been replaced by a class. The reason for this is that this allows debuggers to reference an instance to uniquely identify a source file and it also makes the script available. This Script instance is now also required on breakpoints for them to match. To allow breakpoints to be use this instance, an event has been added to Engine that is called whenever a new Script instance is created.